move date utils to :core:utils
