+++ rss-generator/core/utils/src/main/kotlin/com/livefast/eattrash/rssgenerator/core/utils/di/UtilsModule.kt
@@ -1,0 +1,15 @@
+package com.livefast.eattrash.rssgenerator.core.utils.di
+
+import com.livefast.eattrash.rssgenerator.core.utils.DateUtils
+import com.livefast.eattrash.rssgenerator.core.utils.DateUtilsImpl
+import org.koin.dsl.module
+
+/**
+ * DI module for core utilities.
+ */
+val utilsModule =
+    module {
+        single<DateUtils> {
+            DateUtilsImpl()
+        }
+    }
